refactor: add barrel exports and consolidate dashboard imports
Add index.ts barrel files for api/, hooks/, components/ui/,
components/layout/, charts/, and lib/ directories. Update all
imports across pages, charts, and components to use barrel
exports instead of direct file paths. Reduces import clutter
from 6-10 lines to 3-4 per file.
