+++ dashboard/src/components/layout/index.ts
@@ -1,0 +1,3 @@
+export { Header } from "./header";
+export { Shell } from "./shell";
+export { Sidebar } from "./sidebar";
