+++ dashboard/src/hooks/index.ts
@@ -1,0 +1,9 @@
+export { useApi } from "./use-api";
+export {
+  lastRefreshAt,
+  markRefreshed,
+  refreshInterval,
+  setRefreshInterval,
+} from "./use-auto-refresh";
+export { theme, toggleTheme } from "./use-theme";
+export { addToast, dismissToast, type Toast, toasts } from "./use-toast";
